Rename operator error functions and fix wording in cortex installer
