+++ pkg/operator/endpoints/shared.go
@@ -25,7 +25,7 @@ import (
 	"github.com/cortexlabs/cortex/pkg/api/schema"
 	s "github.com/cortexlabs/cortex/pkg/api/strings"
 	"github.com/cortexlabs/cortex/pkg/lib/errors"
-	oerrors "github.com/cortexlabs/cortex/pkg/operator/errors"
+	"github.com/cortexlabs/cortex/pkg/operator/telemetry"
 )
 
 func Respond(w http.ResponseWriter, response interface{}) {
@@ -39,7 +39,7 @@ func RespondError(w http.ResponseWriter, err error, strs ...string) {
 
 func RespondErrorCode(w http.ResponseWriter, code int, err error, strs ...string) {
 	err = errors.Wrap(err, strs...)
-	oerrors.PrintError(err)
+	errors.PrintError(err)
 
 	w.WriteHeader(code)
 	response := schema.ErrorResponse{
@@ -59,6 +59,7 @@ func RespondIfError(w http.ResponseWriter, err error, strs ...string) bool {
 func RecoverAndRespond(w http.ResponseWriter, strs ...string) {
 	if errInterface := recover(); errInterface != nil {
 		err := errors.CastRecoverError(errInterface, strs...)
+		telemetry.ReportError(err)
 		RespondError(w, err)
 	}
 }
